Httpd reload is moved from handlers to tasks of 170-setup_httpd_RedHat.yml.
A handler is fragile, because it is run at the end of playbook and it will not be called when the playbook exits for some error.  The second run of the playbook might not send a notification to the handler, because the configuration is already changed before the last error.  Thus, reloading httpd in a task is more certain to be called.