This opens httpd_port both in the current state and the config of firewalld.
Previously, the current state of firewalld is updated by the "reload" handler.  However, this does not work when the new open port is added to firewalld ("restarted" in place of "reloaded" would have worked fine).  The new way is not to reload the config file at all, but is to let firewalld module to change the runtime state with permanent=False.